// frontend/src/apiCodegen/api.ts
export interface Info {
    /**
     * 
     * @type {string}
     * @memberof Info
     */
    'message': string;
}
/**
 * 
 * @export
 * @interface ModelError
 */
export interface ModelError {
    /**
     * 
     * @type {string}
     * @memberof ModelError
     */
    'message': string;
}
/**
 * 
 * @export
 * @interface Schema
 */
export interface Schema {
    /**
     * 
     * @type {string}
     * @memberof Schema
     */
    'column': string;
    /**
     * 
     * @type {string}
     * @memberof Schema
     */
    'type': SchemaTypeEnum;
}

export const SchemaTypeEnum = {
    Integer: 'integer',
    Real: 'real',
    Char: 'char',
    String: 'string',
    RealInv: 'realInv',
    Image: 'image'
} as const;

export type SchemaTypeEnum = typeof SchemaTypeEnum[keyof typeof SchemaTypeEnum];